mcp-server/pyproject.toml
[project]
name = "mcp_bsl_server"
version = "0.1.0"
description = "MCP server for querying data via BSL (Boring Semantic Layer)"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.1",
    "pandas",
    "python-dotenv",
    "pyyaml>=6.0.2",
    "dlt[duckdb]>=1.0.0",
    "boring-semantic-layer>=0.2.0",
    "ibis-framework[duckdb]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389"]

[project.scripts]
mcp_bsl_server = "mcp_bsl_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_bsl_server"]
